Fixed zero-byte error
The issue was that autohotkey did not recognize the ahk script
as an utf-8 encoded file since there was no BOM at the beginning
of the script. This has been inserted now.

Also, autohotkey in Unicode mode represents all strings as
utf-16le encoded. That's where the zero byte comes from (none of
the characters I've tried so far have been beyond the first byte
sector so the second byte has always been zero). Correspondingly,
we now decode the string as utf-16le in Python.